Support round robin connection distribution
This factors a worker supervisor out of the spawn worker method of a
cluster supervisor.
The worker supervisor will gain further responsibilities.

This removes usage of the existing, broken in v0.10, Node.js cluster
module.
The first responsibility of the new worker supervisor is to subsume the
responsibility for managing a child process, receiving requests to
accept connections, and forwarding those connections.

This change also introduces a round robin load balancer which manages
the pool of workers that are listening for connections on a particular
port or pipe.
The load balancer accepts connections on behalf of the worker and then
forwards each to one of the workers in turn.
Workers may enter and exit the ring as they go up and down.

Adds two minimal tests for cluster related behaviors, particularly round
robin over a plain socket server, and a verification that the HTTP stack
benefits from the subverted socket layer.