Smooth out cases encountered in practice
handleStop does in fact occur during standby, indicating repeated exit
messages queuing.

The child process object does need to stick around so we can access the
pid even when we return to standby.

Port should default to 0, after the network thunk converts it to false
for distinguishing it from a pipe name.