Documentation and log audit
Documents all of the options and methods for the cluster supervisor.

Documents all log messages, and adjusts the log levels and messages to
make them more insightful and easier to scan.

All worker messages start with worker, and all load balancer and
supervisor messages start with supervisor.
Log levels have been adjusted to reflect the call to action for each
entry.